MoonStat/requirements.txt
mysql-connector-python==8.0.33
protobuf==3.20.3
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.3
